// frontend/src/lib/generated/api/models/DTOPeticionCrearEjercicio.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Academia API
 * API para gestión de academia con roles de administrador, profesor y alumno
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@academia.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Exercise update data
 * @export
 * @interface DTOPeticionCrearEjercicio
 */
export interface DTOPeticionCrearEjercicio {
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearEjercicio
     */
    nombre: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearEjercicio
     */
    enunciado: string;
    /**
     * 
     * @type {Date}
     * @memberof DTOPeticionCrearEjercicio
     */
    fechaInicioPlazo: Date;
    /**
     * 
     * @type {Date}
     * @memberof DTOPeticionCrearEjercicio
     */
    fechaFinalPlazo: Date;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearEjercicio
     */
    claseId: number;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearEjercicio
     */
    duracionEnHoras?: number;
}

/**
 * Check if a given object implements the DTOPeticionCrearEjercicio interface.
 */
export function instanceOfDTOPeticionCrearEjercicio(value: object): value is DTOPeticionCrearEjercicio {
    if (!('nombre' in value) || value['nombre'] === undefined) return false;
    if (!('enunciado' in value) || value['enunciado'] === undefined) return false;
    if (!('fechaInicioPlazo' in value) || value['fechaInicioPlazo'] === undefined) return false;
    if (!('fechaFinalPlazo' in value) || value['fechaFinalPlazo'] === undefined) return false;
    if (!('claseId' in value) || value['claseId'] === undefined) return false;
    return true;
}

export function DTOPeticionCrearEjercicioFromJSON(json: any): DTOPeticionCrearEjercicio {
    return DTOPeticionCrearEjercicioFromJSONTyped(json, false);
}

export function DTOPeticionCrearEjercicioFromJSONTyped(json: any, ignoreDiscriminator: boolean): DTOPeticionCrearEjercicio {
    if (json == null) {
        return json;
    }
    return {
        
        'nombre': json['nombre'],
        'enunciado': json['enunciado'],
        'fechaInicioPlazo': (new Date(json['fechaInicioPlazo'])),
        'fechaFinalPlazo': (new Date(json['fechaFinalPlazo'])),
        'claseId': json['claseId'],
        'duracionEnHoras': json['duracionEnHoras'] == null ? undefined : json['duracionEnHoras'],
    };
}

export function DTOPeticionCrearEjercicioToJSON(json: any): DTOPeticionCrearEjercicio {
    return DTOPeticionCrearEjercicioToJSONTyped(json, false);
}

export function DTOPeticionCrearEjercicioToJSONTyped(value?: DTOPeticionCrearEjercicio | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'nombre': value['nombre'],
        'enunciado': value['enunciado'],
        'fechaInicioPlazo': ((value['fechaInicioPlazo']).toISOString()),
        'fechaFinalPlazo': ((value['fechaFinalPlazo']).toISOString()),
        'claseId': value['claseId'],
        'duracionEnHoras': value['duracionEnHoras'],
    };
}